Completed aspnettest-ized SPA Tests
Aspnettest-ized the SPA tests for minimal.websharper.spa with the
patterns the iselenium library provides:

With the SpaTests base class, formally explicit Selenium WebDriverWait
is now implicitly used by its Click() and Write() methods which wait
RequestTimeout seconds for the element to appear before trying to
click resp. write throttled (char-by-char)  to it.

The new AssertPoll extension method mirrors NUnit's Assert.That and
is intended for statically referenced model objects on the server.
Unlike WebDriverWait which polls the browser DOM, it polls the model
much quicker whether the assertion becomes true within RequestTimeout
seconds. While polling, the browser has time to change the model state
on the server via WebSharper [Remote] RPCs:

this.AssertPoll(() => this.Content, () => Has.Exactly(2).Items);

Cross-browser tests with this NUnit pattern:

    [TestFixture(typeof(ChromeDriver))]
    [TestFixture(typeof(InternetExplorerDriver))]
    public class WithStaticRemoteTest<TWebDriver> : SpaTest<TWebDriver>
        where TWebDriver : IWebDriver, new()

Prettied up the XML result in the Test-Explorer for failed second-order
GUI tests.

+++ iselenium/SeleniumTest.cs
@@ -19,6 +19,11 @@ namespace iselenium
 #pragma warning restore IDE1006
 
         /// <summary>
+        /// Suppress [SetUp] DeleteAllCookies()
+        /// </summary>
+        public bool KeepCookies { get; set; } = false;
+
+        /// <summary>
         /// Start the browser
         /// </summary>
         [OneTimeSetUp]
@@ -42,8 +47,11 @@ namespace iselenium
         [SetUp]
         public void DeleteAllCookies()
         {
-            Thread.Sleep(1000); // avoid eventual OpenQA.Selenium.NoSuchWindowException
-            this.driver.Manage().Cookies.DeleteAllCookies();
+            if (!KeepCookies)
+            {
+                Thread.Sleep(1000); // avoid eventual OpenQA.Selenium.NoSuchWindowException
+                this.driver.Manage().Cookies.DeleteAllCookies();
+            }
         }
     }
 }
